backend/pyproject.toml
[project]
name = "hackathon-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "tenacity>=9.1.4",
    "uvicorn>=0.45.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]
